Improve reliability of "deploy to Render" button
GitHub sets `Referrer Policy: strict-origin-when-cross-origin`, which interacts oddly with Chrome and our Referrer detection. _Sometimes_ Chrome will only send the origin (as set by the policy) but other times it will send the full referrer.

This appears to be a bug (or feature?) in Chrome - The first visit will only send the origin, but navigating back and clicking again will send the full referrer.

Thankfully the anchor tag supports explicitly setting a `referrerpolicy`, telling the browser to always send the full referrer for this link.